Fix build on Windows

+++ src/utils/line_map.cxx
@@ -21,7 +21,7 @@ PLineMap::get_lineno_and_colno(uint32_t p_position, uint32_t* p_lineno, uint32_t
     return;
   } else if (p_position >= m_positions.back()) {
     if (p_lineno)
-      *p_lineno = m_positions.size() + 1;
+      *p_lineno = static_cast<uint32_t>(m_positions.size() + 1);
     if (p_colno)
       *p_colno = p_position - m_positions.back() + 1;
     return;
@@ -66,7 +66,7 @@ uint32_t
 PLineMap::search_rightmost(uint32_t p_position) const
 {
   uint32_t left = 0;
-  uint32_t right = m_positions.size();
+  uint32_t right = static_cast<uint32_t>(m_positions.size());
 
   while (left < right) {
     const uint32_t middle = (left + right) / 2;
